TransactionRepositoryAdapter works.

+++ src/main/java/com/twenty9ine/frauddetection/domain/valueobject/RuleEvaluation.java
@@ -3,6 +3,6 @@ package com.twenty9ine.frauddetection.domain.valueobject;
 import lombok.Builder;
 
 @Builder
-public record RuleEvaluation(String ruleId, String ruleName, RuleType ruleType, boolean triggered, int scoreImpact,
+public record RuleEvaluation(String ruleId, String ruleName, RuleType ruleType, boolean triggered, double scoreImpact,
                              String description) {
 }
